+++ packages/quickjs-emscripten-core/src/context.ts
@@ -1281,26 +1281,35 @@ export class QuickJSContext
       const cause = result.error.consume((error) => this.dump(error))
 
       if (cause && typeof cause === "object" && typeof cause.message === "string") {
-        const { message, name, stack } = cause
-        const exception = new QuickJSUnwrapError("")
-        const hostStack = exception.stack
+        const { message, name, stack, ...rest } = cause
+        const exception = new QuickJSUnwrapError(cause, context)
 
         if (typeof name === "string") {
           exception.name = cause.name
         }
 
+        exception.message = message
+        const hostStack = exception.stack
         if (typeof stack === "string") {
           exception.stack = `${name}: ${message}\n${cause.stack}Host: ${hostStack}`
         }
 
-        Object.assign(exception, { cause, context, message })
+        Object.assign(exception, rest)
         throw exception
       }
 
-      throw new QuickJSUnwrapError(cause, context)
+      throw new QuickJSUnwrapError(cause)
     }
 
     return result.value
+  }
+
+  /** @private */
+  [Symbol.for("nodejs.util.inspect.custom")]() {
+    if (!this.alive) {
+      return `${this.constructor.name} { disposed }`
+    }
+    return `${this.constructor.name} { ctx: ${this.ctx.value} rt: ${this.rt.value} }`
   }
 
   /** @private */