+++ packages/quickjs-emscripten-core/src/errors.ts
@@ -9,7 +9,11 @@ export class QuickJSUnwrapError extends Error {
     public cause: unknown,
     public context?: QuickJSContext,
   ) {
-    super(String(cause))
+    const message =
+      typeof cause === "object" && cause && "message" in cause
+        ? String(cause.message)
+        : String(cause)
+    super(message)
   }
 }