+++ packages/variant-quickjs-asmjs-mjs-release-sync/tsup.config.ts
@@ -1,0 +1,7 @@
+import { extendConfig } from "@jitl/tsconfig/tsup.base.config.js"
+export default extendConfig({
+  entry: ["src/index.ts"],
+  external: [],
+  format: ["esm"],
+  clean: false,
+})