+++ packages/variant-quickjs-singlefile-cjs-release-asyncify/tsup.config.ts
@@ -1,8 +1,7 @@
 import { extendConfig } from "@jitl/tsconfig/tsup.base.config.js"
-
 export default extendConfig({
   entry: ["src/index.ts", "src/ffi.ts"],
-  external: ["@jitl/quickjs-singlefile-cjs-release-asyncify/emscripten-module"],
-  formats: ["cjs"],
+  external: [],
+  format: ["cjs"],
   clean: false,
 })